# openapi.yaml
openapi: 3.0.0
info:
  title: educultura API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3018/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /instructor:
    get:
      tags:
        - Instructor
      summary: Listar todos los Instructors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Instructors
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Instructor
      summary: Crear Instructor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Instructor creado
  /instructor/{id}:
    get:
      tags:
        - Instructor
      summary: Obtener Instructor por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instructor encontrado
        '404':
          description: Instructor no encontrado
    patch:
      tags:
        - Instructor
      summary: Actualizar Instructor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Instructor actualizado
    delete:
      tags:
        - Instructor
      summary: Eliminar Instructor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instructor eliminado
  /instructor/{id}/upload-image:
    post:
      tags:
        - Instructor
      summary: Subir imagen a Instructor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...
  /inscripcion:
    get:
      tags:
        - Inscripcion
      summary: Listar todos los Inscripcions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Inscripcions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Inscripcion
      summary: Crear Inscripcion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Inscripcion creado
  /inscripcion/{id}:
    get:
      tags:
        - Inscripcion
      summary: Obtener Inscripcion por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inscripcion encontrado
        '404':
          description: Inscripcion no encontrado
    patch:
      tags:
        - Inscripcion
      summary: Actualizar Inscripcion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Inscripcion actualizado
    delete:
      tags:
        - Inscripcion
      summary: Eliminar Inscripcion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inscripcion eliminado
  /inscripcion/{id}/upload-image:
    post:
      tags:
        - Inscripcion
      summary: Subir imagen a Inscripcion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...